+++ reduct/bucket.py
@@ -1,7 +1,7 @@
 """Bucket API"""
-from contextlib import asynccontextmanager
 import json
 import time
+from contextlib import asynccontextmanager
 from dataclasses import dataclass
 from enum import Enum
 from typing import (
@@ -11,6 +11,7 @@ from typing import (
     Union,
     Callable,
     Awaitable,
+    Dict,
 )
 
 from pydantic import BaseModel
@@ -110,6 +111,12 @@ class Record:
     read: Callable[[int], AsyncIterator[bytes]]
     """read data in chunks"""
 
+    labels: Dict[str, str]
+    """labels of record"""
+
+
+LABEL_PREFIX = "x-reduct-label-"
+
 
 class Bucket:
     """A bucket of data in Reduct Storage"""
@@ -188,16 +195,7 @@ class Bucket:
         async with self._http.request(
             "GET", f"/b/{self.name}/{entry_name}", params=params
         ) as resp:
-            timestamp = int(resp.headers["x-reduct-time"])
-            size = int(resp.headers["content-length"])
-
-            yield Record(
-                timestamp=timestamp,
-                size=size,
-                last=True,
-                read_all=resp.read,
-                read=resp.content.iter_chunked,
-            )
+            yield self._parse_record(resp)
 
     async def write(
         self,
@@ -205,6 +203,7 @@ class Bucket:
         data: Union[bytes, AsyncIterator[bytes]],
         timestamp: Optional[int] = None,
         content_length: Optional[int] = None,
+        **kwargs,
     ):
         """
         Write a record to entry
@@ -215,6 +214,8 @@ class Bucket:
             timestamp: UNIX time stamp in microseconds. Current time if it's None
             content_length: content size in bytes,
                 needed only when the data is an iterator
+        Keyword Args:
+            labels (dict): labels as key-values
         Raises:
             ReductError: if there is an HTTP error
 
@@ -231,13 +232,14 @@ class Bucket:
 
         """
         params = {"ts": timestamp if timestamp else time.time_ns() / 1000}
-
+        labels = kwargs["labels"] if "labels" in kwargs else None
         await self._http.request_all(
             "POST",
             f"/b/{self.name}/{entry_name}",
             params=params,
             data=data,
             content_length=content_length if content_length is not None else len(data),
+            labels=labels,
         )
 
     async def query(
@@ -246,6 +248,7 @@ class Bucket:
         start: Optional[int] = None,
         stop: Optional[int] = None,
         ttl: Optional[int] = None,
+        **kwargs,
     ) -> AsyncIterator[Record]:
         """
         Query data for a time interval
@@ -255,7 +258,9 @@ class Bucket:
             start: the beginning of the time interval
             stop: the end of the time interval
             ttl: Time To Live of the request in seconds
-
+        Keyword Args:
+            include (dict): query records which have all labels from this dict
+            exclude (dict): querz records which doesn't have all labels from this dict
         Returns:
              AsyncIterator[Record]: iterator to the records
 
@@ -273,6 +278,12 @@ class Bucket:
             params["stop"] = stop
         if ttl:
             params["ttl"] = ttl
+        if "include" in kwargs:
+            for name, value in kwargs["include"].items():
+                params[f"include-{name}"] = str(value)
+        if "exclude" in kwargs:
+            for name, value in kwargs["exclude"].items():
+                params[f"exclude-{name}"] = str(value)
 
         url = f"/b/{self.name}/{entry_name}"
         data = await self._http.request_all(
@@ -286,18 +297,8 @@ class Bucket:
             async with self._http.request("GET", f"{url}?q={query_id}") as resp:
                 if resp.status == 202:
                     return
-
-                timestamp = int(resp.headers["x-reduct-time"])
-                size = int(resp.headers["content-length"])
                 last = int(resp.headers["x-reduct-last"]) != 0
-
-                yield Record(
-                    timestamp=timestamp,
-                    size=size,
-                    last=last,
-                    read_all=resp.read,
-                    read=resp.content.iter_chunked,
-                )
+                yield self._parse_record(resp, last)
 
     async def get_full_info(self) -> BucketFullInfo:
         """
@@ -305,4 +306,21 @@ class Bucket:
         """
         return BucketFullInfo.parse_raw(
             await self._http.request_all("GET", f"/b/{self.name}")
+        )
+
+    def _parse_record(self, resp, last=True):
+        timestamp = int(resp.headers["x-reduct-time"])
+        size = int(resp.headers["content-length"])
+        labels = dict(
+            (name[len(LABEL_PREFIX) :], value)
+            for name, value in resp.headers.items()
+            if name.startswith(LABEL_PREFIX)
+        )
+        return Record(
+            timestamp=timestamp,
+            size=size,
+            last=last,
+            read_all=resp.read,
+            read=resp.content.iter_chunked,
+            labels=labels,
         )